try to build all std dependencies including std
this makes Xargo robust against changes in the std facade as all the
crate names that were hardcoded in Xargo, e.g. `rustc_unicode`, have
been removed.

The downside is that this also worsens the user experience because when
Xargo fails to build one of std dependencies it will simply regurgitate
the error to stderr.