allow "multi-stage" sysroot builds
with this is now possible to build the `test` crate which doesn't
explicitly depend on `std`.

This Xargo.toml

``` toml
[dependencies.std]
features = ["panic_unwind"]
# stage = 0  # implicit

[dependencies.test]
stage = 1
```

Builds the sysroot in two stages. First, Xargo builds the `std`
facade, the stage 0, turns that into a partial sysroot and then builds
the `test` crate, the stage 1, against that partial sysroot. Finally,
the stage 0 artifacts and the stage 1 artifacts are then merged into the
sysroot that will be used to build the application.

The result is that you can use `xargo test` with this Xargo.toml.

closes #104