Use RUSTC_BOOTSTRAP to build sysroot for stable
When building the sysroot or rust compiler with a beta or stable rustc
binary the flag RUSTC_BOOTSTRAP can be set to 1 to allow rustc to allow
the same features as the development channel. This is required as the
rust compiler and sysroot make use of unstable features.

See here: https://github.com/rust-lang/cargo/pull/5613